+++ CHANGES.md
@@ -1,4 +1,5 @@
 #### [unreleased]
+* show overridden plugins/themes by [Skip Updates](https://wordpress.org/plugins/skip-updates/) plugin in Git Updater Settings tab
 
 #### 10.6.14 / 2022-02-05
 * allow hooks to run if no settings to be saved in `Settings::update_settings()`